renamed RobotArmInterface to RobotPanel
