+++ src/main/java/com/marginallyclever/robotoverlord/components/RobotComponent.java
@@ -6,8 +6,8 @@ import com.marginallyclever.robotoverlord.Entity;
 import com.marginallyclever.robotoverlord.RobotOverlord;
 import com.marginallyclever.robotoverlord.robots.Robot;
 import com.marginallyclever.robotoverlord.components.robot.robotarm.ApproximateJacobian2;
-import com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.DHTable;
-import com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.RobotArmInterface;
+import com.marginallyclever.robotoverlord.components.robot.robotarm.robotpanel.DHTable;
+import com.marginallyclever.robotoverlord.components.robot.robotarm.robotpanel.RobotPanel;
 import com.marginallyclever.robotoverlord.swinginterface.view.ViewElementButton;
 import com.marginallyclever.robotoverlord.swinginterface.view.ViewPanel;
 
@@ -55,7 +55,7 @@ public class RobotComponent extends Component implements Robot {
                 try {
                     JDialog frame = new JDialog(parentFrame, "Control panel");
                     frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-                    frame.add(new RobotArmInterface(me));
+                    frame.add(new RobotPanel(me));
                     frame.pack();
                     frame.setLocationRelativeTo(parentFrame);
                     frame.setVisible(true);
