+++ src/main/java/com/marginallyclever/robotoverlord/components/robot/robotarm/robotpanel/ClassExplorer.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface;
+package com.marginallyclever.robotoverlord.components.robot.robotarm.robotpanel;
 
 import com.marginallyclever.communications.application.TextInterfaceToListeners;
 import com.marginallyclever.communications.application.TextInterfaceToSessionLayer;
