+++ src/main/java/com/marginallyclever/robotoverlord/components/robot/robotarm/robotpanel/jogpanel/JogPanel.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.jogpanel;
+package com.marginallyclever.robotoverlord.components.robot.robotarm.robotpanel.jogpanel;
 
 import com.marginallyclever.convenience.log.Log;
 import com.marginallyclever.robotoverlord.components.RobotComponent;
