+++ src/main/java/com/marginallyclever/robotoverlord/components/robot/robotarm/robotpanel/programpanel/programPanel.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.programpanel;
+package com.marginallyclever.robotoverlord.components.robot.robotarm.robotpanel.programpanel;
 
 import com.marginallyclever.convenience.log.Log;
 import com.marginallyclever.robotoverlord.components.RobotComponent;
